An initial stab at RDT via go-back-n.
This implements the ACK dance that the client and server will need to do. It also implements packet timeout to detect failures and resend packets. Note that we only use timeout detection--no duplicate ACK checking--for simplicity.

This code functions with @ipetkov's original test code. Still to do is send data larger than a single packet.